quicly_send not emitting a packet upon timeout is a non-issue
We used to assert that `quicly_send` is called only when there is
something to send.  However the loss detection logic (implemented
based on the recovery I-D) does not guarantee that; when the
sender is in recovery, RTO timers can be triggered without the
possibility to resend anything.

This is now detected due to our exchangig more lower epoch
packets. Hence the relaxation of the too-strict check.