Recognize ACK for a packet sent upon loss without any data.
Quicly sends a packet that only contains an empty CRYPTO frame or a
PADDING frame when it sees a loss event and there is nothing to
send. However, in case of sending a PADDING frame, it is not
registering the on_ack handler for the CRYPTO frame (see
`_do_prepare_packet`).

That lead to the ACK for that sent packet being ignored, because
quicly has assumed the existence of on_ack_stream callback in the
inflight ACK map for legitimate ACKs against Initial, Handshake
packets (see `ack_is_handshake_flow` that is removed by this
commit).

This commit fixes the issue by introducing `ack_epoch` as a member
of `quicly_ack_t`.