Cut out API that allows h2o to have it's own reassembly buffer.
Previously, quicly has provided `quicly_buffer_t` that was used as
the send / receive buffer.  The design was inefficient considering
the fact that H2O has it's own buffers that is used for holding
inflight data (e.g., unsent data, partially-received H3 frames).

This PR removes the redundancy by introducing an API
(`quicly_stream_callbacks_t`) that allows H2O to supply callbacks
for filling the recieve buffer / writing directly to packet
payload.

We also have a new simple buffer called `quicly_streambuf_t` as
part of quicly.  It is used for buffering crypto streams as well
as in tests.