Do not update sentmap, once the connection enters CLOSING state.
When entering CLOSING or DRAINING state, quicly clears the sentmap
and registers a sentry. Then the connection is closed when the
number of entries in sentmap goes down to zero, as that sentry gets
evicted after 4 PTO.

In PR #256, we started responding to incoming packets in CLOSING
state, by sending a new packet. However, this emission has inserted
new entries to the sentmap, and the connection stays in the CLOSING
state continues longer than 4 PTO.