Use pre-allocated storage for NCID/RCID queues
This commit changes the way quicly internally stores NEW_CONNECTION_ID
and RETIRE_CONNECTION_ID frames to be sent in the next send round.
Instead of dynamically allocating an object for each frame, this commit
uses fixed-length array to limit resource usage for these frames.

Note: it is possible that some RETIRE_CONNECTION_ID frame might not
be able to find a slot to be stored in a very rare case, if the
retirement happens to fast. In such a scenario we let the RETIRE_CONNECTION_ID
frame lost i.e. we would retire it internally but the peer would
never know it.

Along with the changes above, this patch also changes when connection
ID and stateless reset token are computed. Now we compute them once
we schedule a NEW_CONNECTION_ID frame for the first time, instead of
each time we send it over wire. CID encryptor might change over time,
therefore recomputing CID/stateless reset token every time might end
up with generating different CID/token upon retransmission.