Lift assumption on NCID receive order
Previous implementation of NEW_CONNECTION_ID had an assumption
that connection IDs are received in the order of sequence number,
and that a connection ID with a smaller sequence than the currently
active ones was considered already retired.

This patch removes such an assumption, by explicitly enumerating
sequence numbers that are expected to arrive.
If the received sequence number exceeds largest_sequence_expected,
it is an ACTIVE_CONNECTION_LIMIT error.
If the number is not found in the list of expected sequence numbers,
it is considered to be already retired.