Add module to manage issued CID logic
This is the first step to separate NEW_CONNECTION_ID issuing logic
from network transport code.

Along with refactoring, this patch introduces a change in logic
that now pending CIDs (with state==STATE_PENDING) are placed at the
front of the CID array in FIFO manner. This will simplify the code
when sending frames, as well as preserve the order of frames as
they were generated.