Add module to encapsulate retire CID issuing logic
This is the first step to refactor RETIRE_CONNECTION_ID handling
logic. This patch introduces a module to encapsulate the state
handling logic.

Along with refactoring, this patch also introduces a change in logic
that now pending sequences are packed at the front of the array.
This will make sending logic simpler and keep the order of frames
in which they were generated.