Reroute partial send path for NCID/RCID
This is a refactoring patch to change the flow in NCID/RCID partial
send path so that the update function invocations (issued_cid_on_sent
and retire_cid_pop) are consolidated in one place.
This will make the intention of the code (update is done for multiple
CIDs at once) much clearer.