Suppress excessive emission of ack-eliciting packets that lead to ping-pong when the MTU is below 1200 bytes.
Up until now, quicly running as a client has always been bundling ack-eliciting
frames to an ACK, until it learns that the server has validated the client's
address.

With this commit, ack-eliciting frames (i.e. CRYPTO data or PING) is sent only
when the PTO timer fires. When running as a server, the ack-eliciting frames
will be sent in both Initial and Handshake packets (if the keys are available).
When running as a client, the ack-eliciting frames will be sent in one of those
two packet types (with Handshake packet being the precedence).