Eliminate `is_active` flag, as it is fragile to bugs.
The flag is only usable when there is a guarantee that there would
not be more than one inflight sentmap carrying the same byte. But
that does not stand for the FIN flag, causing assertion failures.
`quicly_ranges_subtract` is going to be fast unless packet loss is
involved (due to end offset being acked being no greater than the
beginning offset of pending ranges), therefore this optimization is
unnecessary.